requirements.txt
aiogram==3.25.0
python-dotenv==1.2.1
gigachat==0.2.0
gspread_asyncio==2.0.0
oauth2client==4.1.3